v1/devel/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]

name = "tincan"
dynamic = ["version"]
description = "Tincan Network Management System"
readme = "README.rst"
requires-python= ">= 3.9"
license = { file = "LICENSE" }
authors = [
    { name= 'Todd Valentic', email='todd.valentic@gmail.com' }
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Public License v3 (GPLv3)",
    "Operating System :: OS Independent"
]

dependencies = [
    "pytz",
    "dateutils",
    "eventlet",
    "requests",
    "bcrypt",
    "flask",
    "psycopg2-binary",
    "sqlalchemy",
    "Flask-SQLAlchemy",
    "flask-compress",
    "flask-script",
    "flask-migrate",
    "flask-bcrypt",
    "flask-jwt-extended",
    "flask-socketio",
    "flask-restful",
    "flask-mail",
    "flask-flatpages",
    "coverage",
    "PyJWT",
    "cryptography",
    "itsdangerous",
    "docutils"
]

[project.optional-dependencies]

test = [
    "pytest",
    "pyfakefs"
    ]

devel = [
    "build",
    "pylint",
    "black"
    ]

[project.urls]

homepage="https://github.com/valentic/tincan_network"
repository="https://github.com/valentic/tincan_network"

[project.scripts]

server-cmd = "server.commands:cli"
server-launch = "server.commands:launch"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "server.metadata.version"}

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib"
]

